refactor: move store.ts to shared/store.ts
