+++ packages/base/src/formbuilder/objectform/source/wmsTileSource.tsx
@@ -10,7 +10,7 @@ import {
 } from '@/src/formbuilder/objectform/schemaUtils';
 import type { ISourceFormProps } from '@/src/formbuilder/objectform/source/sourceform';
 import { useSchemaFormState } from '@/src/formbuilder/objectform/useSchemaFormState';
-import { GlobalStateDbManager } from '@/src/store';
+import { GlobalStateDbManager } from '@/src/shared/store';
 import { deepCopy } from '@/src/tools';
 import { IWmsLayerInfo } from '@/src/types';
